backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
aiosqlite==0.21.0
greenlet==3.3.0
aiogram==3.3.0
google-generativeai==0.3.2
python-dotenv==1.0.1
streamlit
pandas
pydantic==2.5.3
pydantic-settings==2.1.0
requests==2.31.0
beautifulsoup4==4.12.3
